fix(git): address pathspec review findings
